Move audio.Synthesizer to separate file
Because audio.go is big enough.
