+++ audio/audio.go
@@ -7,13 +7,6 @@
 // Sound effects and music can be changed using functions: SetSfx and SetMusic.
 package audio
 
-import (
-	"bytes"
-	"fmt"
-
-	"github.com/elgopher/pi"
-)
-
 // Sfx starts playing sound effect with given sfxNo on specified channel.
 //
 // sfxNo is the number of sound effect to play (0-63). sfxNo=-1 stops
@@ -258,239 +251,6 @@ type Pattern struct {
 type PatternSfx struct {
 	SfxNo   byte // 0-63
 	Enabled bool
-}
-
-// Synthesizer is used by a back-end. It is an AudioSystem implementation.
-// Plus it provides method for synthesizing audio samples.
-type Synthesizer struct {
-	sfx     map[byte]SoundEffect
-	pattern map[byte]Pattern
-}
-
-// ReadSamples method is used by a back-end to read generated audio samples and play them back to the user. The sample rate is 22050.
-// One channel is used (mono).
-//
-// ReadSamples is (usually) executed concurrently with main game loop. Back-end should add proper synchronization to avoid race conditions.
-// Back-end could decide about buffer size, although the higher the size the higher the lag. Usually the buffer is 512 samples,
-// which is 23ms of audio.
-func (s *Synthesizer) ReadSamples(b []float64) (n int, err error) {
-	if len(b) == 0 {
-		return 0, nil
-	}
-
-	// generate silence for now
-	for i := 0; i < len(b); i++ {
-		b[i] = 0
-	}
-
-	return len(b), nil
-}
-
-func (s *Synthesizer) Sfx(sfxNo int, channel Channel, offset, length int) {
-	fmt.Println("Sfx is not implemented yet. Sorry...")
-}
-
-func (s *Synthesizer) Music(patterNo int, fadeMs int, channelMask byte) {
-	fmt.Println("Music is not implemented yet. Sorry...")
-}
-
-func (s *Synthesizer) Stat() Stat {
-	fmt.Println("Stat is not implemented yet. Sorry...")
-	return Stat{}
-}
-
-const maxSfxNo = 63
-
-func (s *Synthesizer) SetSfx(sfxNo int, effect SoundEffect) {
-	if sfxNo < 0 || sfxNo > maxSfxNo {
-		return
-	}
-
-	if s.sfx == nil {
-		s.sfx = map[byte]SoundEffect{}
-	}
-
-	effect.LoopStart = pi.MinInt(63, effect.LoopStart)
-	effect.LoopStop = pi.MinInt(63, effect.LoopStop)
-	effect.Detune = pi.MinInt(2, effect.Detune)
-	effect.Reverb = pi.MinInt(2, effect.Reverb)
-	effect.Dampen = pi.MinInt(2, effect.Dampen)
-
-	for i := 0; i < len(effect.Notes); i++ {
-		volume := effect.Notes[i].Volume
-		effect.Notes[i].Volume = pi.MinInt(7, volume)
-
-		pitch := effect.Notes[i].Pitch
-		effect.Notes[i].Pitch = pi.MinInt(63, pitch)
-
-		instrument := effect.Notes[i].Instrument
-		effect.Notes[i].Instrument = pi.MinInt(15, instrument)
-
-		eff := effect.Notes[i].Effect
-		effect.Notes[i].Effect = pi.MinInt(7, eff)
-	}
-
-	s.sfx[byte(sfxNo)] = effect
-}
-
-func (s *Synthesizer) GetSfx(sfxNo int) (e SoundEffect) {
-	if sfxNo < 0 || sfxNo > maxSfxNo {
-		return e
-	}
-
-	return s.sfx[byte(sfxNo)]
-}
-
-const maxPatternNo = 63
-
-func (s *Synthesizer) SetMusic(patternNo int, pattern Pattern) {
-	if patternNo < 0 || patternNo > maxPatternNo {
-		return
-	}
-
-	if s.pattern == nil {
-		s.pattern = map[byte]Pattern{}
-	}
-
-	for i := 0; i < len(pattern.Sfx); i++ {
-		sfxNo := pattern.Sfx[i].SfxNo
-		pattern.Sfx[i].SfxNo = pi.MinInt(maxSfxNo, sfxNo)
-	}
-
-	s.pattern[byte(patternNo)] = pattern
-}
-
-func (s *Synthesizer) GetMusic(patterNo int) (p Pattern) {
-	if patterNo < 0 || patterNo > maxPatternNo {
-		return p
-	}
-
-	return s.pattern[byte(patterNo)]
-}
-
-const schemaVersion = 1
-
-func (s *Synthesizer) Load(state []byte) error {
-	if len(state) == 0 {
-		return fmt.Errorf("state is empty")
-	}
-
-	version := state[0]
-	if schemaVersion != version {
-		return fmt.Errorf("state version %d is not supported. Only %d is supported.", version, schemaVersion)
-	}
-
-	const expectedStateLen = 9409
-	if len(state) != expectedStateLen {
-		return fmt.Errorf("invalid length of state. Must be %d.", expectedStateLen)
-	}
-
-	offset := 1
-
-	for sfxNo := 0; sfxNo <= maxSfxNo; sfxNo++ {
-		var sfx SoundEffect
-
-		for j, note := range sfx.Notes {
-			note.Pitch = Pitch(state[offset])
-			offset++
-			note.Instrument = Instrument(state[offset])
-			offset++
-			note.Volume = Volume(state[offset])
-			offset++
-			note.Effect = Effect(state[offset])
-			offset++
-
-			sfx.Notes[j] = note
-		}
-
-		sfx.Speed = state[offset]
-		offset++
-		sfx.LoopStart = state[offset]
-		offset++
-		sfx.LoopStop = state[offset]
-		offset++
-		sfx.Detune = state[offset]
-		offset++
-		sfx.Reverb = state[offset]
-		offset++
-		sfx.Dampen = state[offset]
-		offset++
-		sfx.Noiz = byteToBool(state[offset])
-		offset++
-		sfx.Buzz = byteToBool(state[offset])
-		offset++
-
-		s.SetSfx(sfxNo, sfx)
-	}
-
-	for patterNo := 0; patterNo <= maxPatternNo; patterNo++ {
-		var pattern Pattern
-
-		for j, sfx := range pattern.Sfx {
-			sfx.SfxNo = state[offset]
-			offset++
-			sfx.Enabled = byteToBool(state[offset])
-			offset++
-
-			pattern.Sfx[j] = sfx
-		}
-
-		pattern.BeginLoop = byteToBool(state[offset])
-		offset++
-		pattern.EndLoop = byteToBool(state[offset])
-		offset++
-		pattern.StopAtTheEnd = byteToBool(state[offset])
-		offset++
-
-		s.SetMusic(patterNo, pattern)
-	}
-
-	return nil
-}
-
-func (s *Synthesizer) Save() ([]byte, error) {
-	buffer := bytes.NewBuffer(nil)
-
-	buffer.WriteByte(schemaVersion)
-
-	for i := 0; i <= maxSfxNo; i++ {
-		sfx := s.GetSfx(i)
-		for _, note := range sfx.Notes {
-			buffer.WriteByte(byte(note.Pitch))
-			buffer.WriteByte(byte(note.Instrument))
-			buffer.WriteByte(byte(note.Volume))
-			buffer.WriteByte(byte(note.Effect))
-		}
-		buffer.WriteByte(sfx.Speed)
-		buffer.WriteByte(sfx.LoopStart)
-		buffer.WriteByte(sfx.LoopStop)
-		buffer.WriteByte(sfx.Detune)
-		buffer.WriteByte(sfx.Reverb)
-		buffer.WriteByte(sfx.Dampen)
-		buffer.WriteByte(boolToByte(sfx.Noiz))
-		buffer.WriteByte(boolToByte(sfx.Buzz))
-	}
-
-	for i := 0; i <= maxPatternNo; i++ {
-		pattern := s.GetMusic(i)
-		for _, sfx := range pattern.Sfx {
-			buffer.WriteByte(sfx.SfxNo)
-			buffer.WriteByte(boolToByte(sfx.Enabled))
-		}
-		buffer.WriteByte(boolToByte(pattern.BeginLoop))
-		buffer.WriteByte(boolToByte(pattern.EndLoop))
-		buffer.WriteByte(boolToByte(pattern.StopAtTheEnd))
-	}
-
-	return buffer.Bytes(), nil
-}
-
-func boolToByte(b bool) byte {
-	if b {
-		return 1
-	}
-
-	return 0
 }
 
 func byteToBool(b byte) bool {
