+++ audio/synth_bench_test.go
